refactor: Rename company-related files to use singular form
